refactor: part 属性/自动化轨字段全量归一化为小数刻度（跨引擎复用契约）
宿主切换引擎/音源不清空 part 已存属性与 automation 曲线，同 key 会被静默
复用。统一字段刻度为最通用的归一化小数（1.0=原样/满刻度），贴合公共认知、
撞键也不越界。全部行为保持——合成结果不变，换算系数同步抵消。

- gender [-100,100]→[-1,1]（GenderConvert 去 /100）
- speed  [0,200]/100→[0,2]/1（SpeedConvert 改 2^(x-1)）
- energy/breathiness/voicing/tension 编辑轨 [-100,100]/[0,100]→[-1,1]/[0,1]
  （Delta 系数 ×100；回显轨保持真实 dB——不入 part、无复用风险）
- speaker mix 权重 [0,100]→[0,1]（SpeakerMix 去 ×0.01）
- seed_* [0,65536]→[0,1]，合成期 round(v·uint.MaxValue) 放大到 uint32
  喂哈希（哈希白化→刻度不影响质量）；seedPerFrame int[]→uint[] 贯通
  Noise/Pitch/Variance；仅精确 v=0 保留 take-0

契约写入 docs/tunelab-voicebank-schema.md §14（key=不可变契约、逐字段量程表、
seed 标称语义、language 仅建议 ISO）。

+++ DiffSingerNoise.cs
@@ -19,14 +19,14 @@ public static class DiffSingerNoise
     public const ulong StageAcoustic = 0xA0, StagePitch = 0xB0, StageVariance = 0xC0;
 
     // 标量 seed（全 part 一致，如 acoustic timbre——它全局不局部化，单一 seed 即可）。
-    public static void AddNoise(List<NamedOnnxValue> inputs, InferenceSession model, int seed, ulong stage, int nFrames)
+    public static void AddNoise(List<NamedOnnxValue> inputs, InferenceSession model, uint seed, ulong stage, int nFrames)
         => AddNoise(inputs, model, _ => seed, stage, nFrames);
 
-    // 逐帧 seed（pitch/variance 的 seed 自动化轨：每帧可不同 → 时间维 × 值维 = 区域独立 take）。
-    public static void AddNoise(List<NamedOnnxValue> inputs, InferenceSession model, int[] seedPerFrame, ulong stage, int nFrames)
+    // 逐帧 seed（pitch/variance 的 seed 自动化轨：每帧可不同 → 时间维 × 值维 = 区域独立 take）。归一化 [0,1] 已在会话期放大到 uint32。
+    public static void AddNoise(List<NamedOnnxValue> inputs, InferenceSession model, uint[] seedPerFrame, ulong stage, int nFrames)
         => AddNoise(inputs, model, t => seedPerFrame[Math.Min(t, seedPerFrame.Length - 1)], stage, nFrames);
 
-    static void AddNoise(List<NamedOnnxValue> inputs, InferenceSession model, Func<int, int> seedAt, ulong stage, int nFrames)
+    static void AddNoise(List<NamedOnnxValue> inputs, InferenceSession model, Func<int, uint> seedAt, ulong stage, int nFrames)
     {
         if (!model.InputMetadata.TryGetValue("noise", out var meta))
             return;   // 无 noise 口（stock 模型）→ 不喂，退回模型内部随机
@@ -37,7 +37,7 @@ public static class DiffSingerNoise
         for (int c = 0; c < feats; c++)
             for (int m = 0; m < outDims; m++)
                 for (int t = 0; t < nFrames; t++)
-                    data[idx++] = Gaussian(stage, (ulong)(uint)seedAt(t), (ulong)c, (ulong)m, (ulong)t);
+                    data[idx++] = Gaussian(stage, seedAt(t), (ulong)c, (ulong)m, (ulong)t);
         inputs.Add(NamedOnnxValue.CreateFromTensor("noise",
             new DenseTensor<float>(data, new[] { 1, feats, outDims, nFrames })));
     }
